+++ src/searches.c
@@ -279,8 +279,8 @@ int chemfp_count_tanimoto_arena(
       start_target_popcount = 0;
       end_target_popcount = num_bits;
     } else {
-      start_target_popcount = query_popcount * threshold;
-      end_target_popcount = ceil(query_popcount / threshold);
+      start_target_popcount = (int)(query_popcount * threshold);
+      end_target_popcount   = (int)(ceil(query_popcount / threshold));
       if (end_target_popcount > num_bits) {
         end_target_popcount = num_bits;
       }
@@ -438,8 +438,8 @@ int chemfp_threshold_tanimoto_arena(
       start_target_popcount = 0;
       end_target_popcount = num_bits;
     } else {
-      start_target_popcount = query_popcount * threshold;
-      end_target_popcount = ceil(query_popcount / threshold);
+      start_target_popcount = (int)(query_popcount * threshold);
+      end_target_popcount   = (int)(ceil(query_popcount / threshold));
       if (end_target_popcount > num_bits) {
         end_target_popcount = num_bits;
       }
